+++ index.js
@@ -45,7 +45,10 @@ CondensationTests.prototype.hOptsFunc = function() {
  * @param {String} particlePath Relative path to the particle
  * @param {Object} [options] processing configuration
  * @param {Object} [options.logicalId] the Logical ID for the particle
- * @param {Object} [options.hArgs] Handlebars arguments to use for processing the particle
+ * @param {Object} [options.cOpts] Mock condensation settings for the current test
+ * @param {Array} [options.hArgs] Mock an array of arguments sent by handlebars
+ * @param {Object} [options.hOpts] Mock the options hash passed by handlebars
+ * @param {Object} [options.hashOpts] shortcut for setting {hOpts: {hash: {} }}
  * @param {Boolean} [options.validateJson=true] Whether to valide the output as valid JSON
  * @param {Boolean} [options.expectError=false] If `true` expect the test to throw an error
  * @return {String} The processed particle
@@ -59,9 +62,24 @@ CondensationTests.prototype.processParticle = function(particleType,particlePath
 
   var hOpts = this.hOptsFunc();
   hOpts.hash.logicalId = options.logicalId;
-  hOpts.hash = _.merge(hOpts.hash,options.hArgs);
+  hOpts.hash = _.merge(hOpts.hash, options.hashOpts);
+  hOpts = _.merge(hOpts,options.hOpts);
 
-  var result = this.condensation.helpers[particleType].apply(null,[particlePath, hOpts]);
+  var cOpts = options.cOpts;
+  var hArgs = options.hArgs;
+
+  /*
+   * For Backwards compatibility only
+   * Prior to 0.5.4 hArgs was incorrecty used instead of hOpts
+   *
+   * hArg should be an array of string arguments, but if it is an object then it will be treated like hOpts
+   */
+  if (options.hArgs && !_.isArray(options.hArgs)) {
+    hOpts.hash = _.merge(hOpts.hash,options.hArgs);
+    delete options.hArgs;
+  }
+
+  var result = this.condensation.helpers[particleType].apply(cOpts,_.compact(_.flatten([particlePath, hArgs, hOpts])));
 
   if (options.validateJson) {
     try {
@@ -83,8 +101,11 @@ CondensationTests.prototype.processParticle = function(particleType,particlePath
  * @param {String} particlePath Relative path to the particle
  * @param {Object} fixture A fixtue to compare the processed output to
  * @param {Object} [options] processing configuration
- * @param {Object} [options.logicalId] the Logical ID for the particle
- * @param {Object} [options.hArgs] Handlebars arguments to use for processing the particle
+ * @param {Object} [options.cOpts] Mock condensation settings for the current test
+ * @param {Array} [options.hArgs] Mock an array of arguments sent by handlebars
+ * @param {Object} [options.hOpts] Mock the options hash passed by handlebars
+ * @param {Object} [options.hashOpts] shortcut for setting {hOpts: {hash: {} }}
+ * @param {Object} [options.logicalId] shortcut for {hOpts: {hash: {logicalId: "" }}}
  * @param {Boolean} [options.validateJson=true] Whether to valide the output as valid JSON
  * @param {String} [options.assertType] How to assert against the fixture. If options.validateJson is true then assertType will be deepEqual. If false then equal
  * @return {String} The processed particle